requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
numpy==1.23.0
py-cord==2.3.2
python-dotenv==0.20.0
youtube-dl==2021.12.17
imageio==2.19.3
requests==2.28.1
cryptography==38.0.3
topggpy==1.4.0
dice==3.1.2
#pyttsx3==2.71
gTTS==2.30